LayoutTests/pdf/unifiedpdf/compositing/pdf-scrolling-tree.html

<!DOCTYPE html> <!-- webkit-test-runner [ UnifiedPDFEnabled=true ] -->
<html>
<head>
    <script src="../../../resources/js-test-pre.js"></script>
    <style>
        embed {
            width: 300px;
            height: 100px;
        }
    </style>
    <script src="../../../resources/ui-helper.js"></script>
    <script>
        window.jsTestIsAsync = true;

        if (window.testRunner)
            testRunner.dumpAsText();

        window.addEventListener('load', async () => {
            await UIHelper.renderingUpdate();

            if (!window.internals)
                return;
           
            shouldBecomeEqual("internals.numberOfScrollableAreas()", "1", () => {
                document.getElementById('scrollingtree').textContent = internals.scrollingStateTreeAsText();
                finishJSTest();
            });
        }, false);
    </script>
</head>
<body>
    <embed src="../../../fast/images/resources/green_rectangle.pdf">
    <pre id="scrollingtree"></pre>
    <script src="../../../resources/js-test-post.js"></script>
</body>
</html>
